+++ src/Cmdty.Storage.Excel/IntrinsicXl.cs
@@ -56,11 +56,9 @@ namespace Cmdty.Storage.Excel
                     double numericalTolerance = StorageExcelHelper.DefaultIfExcelEmptyOrMissing(numericalToleranceIn, 1E-10,
                                                                     "Numerical_tolerance");
 
-                    // TODO provide alternative method for interpolating interest rates
-                    Func<Day, double> interpolatedInterestRates =
-                        StorageExcelHelper.CreateLinearInterpolatedInterestRateFunc(interestRateCurve, ExcelArg.InterestRateCurve.Name);
+                    Func<Day, Day, double> discountFunc =
+                        StorageExcelHelper.CreateLogLinearInterpolatedDiscountFactors(interestRateCurve, ExcelArg.InterestRateCurve.Name);
                     Func<Day, Day> settleDateRule = StorageExcelHelper.CreateSettlementRule(settleDatesIn, ExcelArg.SettleDates.Name);
-                    Func<Day, Day, double> discountFunc = StorageHelper.CreateAct65ContCompDiscounter(interpolatedInterestRates);
 
                     var storage = ObjectCache.Instance.GetObject<CmdtyStorage<Day>>(storageHandle);