+++ CHANGELOG.md
@@ -34,6 +34,10 @@ data is returned.
 * Change to allowable values Terminal_inventory argument of cmdty.CreateStorage function and addition of
 associated Terminal_val_param argument.
 
+### 0.3.0 (Not yet released)
+* Async and blocking modes for calculation.
+* Log-linear interpolation of discount factors.
+
 ---
 ## NuGet Package Releases
 ### 1.1.0